+++ source/graphics/3d/DirectDraw.h
@@ -54,8 +54,8 @@ DirectDraw DirectDraw_getInstance();
 
 void DirectDraw_destructor(DirectDraw this);
 void DirectDraw_drawLine(DirectDraw this, VBVec2D fromPoint, VBVec2D toPoint, int color);
-void DirectDraw_drawPixel(DirectDraw this, u32 buffer, int x, int y, int color);
-void DirectDraw_drawBlackPixel(DirectDraw this, u32 buffer, int x, int y);
+void DirectDraw_drawPixel(DirectDraw this, u32 buffer, u16 x, u16 y, int color);
+void DirectDraw_drawBlackPixel(DirectDraw this, u32 buffer, u16 x, u16 y);
 
 
 #endif
